Fixed bug in select-editors init;
Backbone-Forms will only fire 'change' event when options are ajaxed, so
that the selector-plugins will pickup and reRender even if it is
initialized before the options change.
Form views in modules will now have total control on which kinda select
plugin they want to use in the onRenderPlus() extension point.