Working on the admin factory and form widget
The config passed to app.Factory.AdminModule.create() should be the
first entry point for admin module developers. Thus we will not store
this config in admin factory nor the admin modules atm, the admin
modules will not be aware of the configs but to hold and show the
grid/form widgets.