Fixed sub-grid form editor bug.
Due to the newly added show event (editor:show) propagation problem in
modified backbone-forms lib.
It kinda bubbles to the parent form editor (like a sub-grid) and causing
it to re-render itself thus re-fetch the collections and thus messing up
the form model's collection reference. The model that still shown on the
form is deprecated by the time the sub-grid editor done re-fectching
like twice creating new collection and models.