Separated banner into a UI module under Shared.
Added Shared context for modules like banner and footer;
Added data-unit name to all the created models;
Added username display in banner module;
Added Dev Rule:
Submodules of a 'context' will not directly listen to events from a
submodule of the same or another 'context'.  The event must be
fired/heard on the 'context' module.