Merged client lib with the jslib project.
This is to maintain the lib version and combined libs build easier;
Spawning new project using the current client infrustructure and libs is
shifted to the shoulder of /tools/spawn where new project index.html and
required files are prepared to  create a new development project. (Note
that this is quite different than building a project, see below)
Building under-development project for demo and testing or even
distribution is the job of /tools/build, within which the project will
be combined compressed and output-ed according to given build config
file.