Refined gzip shared tool script.
Can auto gzip cached scripts through build/spawn config.
Note that the spawn tool will not minify any of the grouped script, and
the build tool will only group all the scripts and produces both
minified/non-minified version of the all-in-one script.
Gzip or not (instead of minified or not) can be specified to the grouped
scripts through config files.